refactor: 주문 DTO 리팩토링
