+++ src/test/kotlin/kr/hhplus/be/application/service/OrderServiceTest.kt
@@ -7,7 +7,7 @@ import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
-import kr.hhplus.be.application.order.OrderCreateDto
+import kr.hhplus.be.application.order.OrderDto.OrderCreateDto
 import kr.hhplus.be.domain.exception.BusinessException
 import kr.hhplus.be.domain.exception.ErrorCode
 import kr.hhplus.be.domain.order.Order
